requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
duckdb==1.1.3
# yfinance==0.2.48  # Replaced with Alpha Vantage API (uses requests)
pandas==2.2.3
numpy==2.1.3
requests==2.32.3
vercel==0.3.2
vercel-sandbox==0.0.2
vercel-sdk==0.0.8
